Stage now looks good while resizing

+++ src/main/java/DataManagement/SeriesCollector.java
@@ -24,7 +24,7 @@ public class SeriesCollector {
             while((line = reader.readLine()) != null){
                 String[] values = line.split(";");
                 for (int i = 0; i < names.length-1 ; i++) {
-                    Series[i].getData().add(new XYChart.Data<String,Number>(values[0],Integer.valueOf(values[i+1])));
+                    Series[i].getData().add(new XYChart.Data<String,Number>(values[0],Double.valueOf(values[i+1])));
                 }
             }
             reader.close();
